Collect EMTs and NPLs using LightSamplerBase
