+++ src/appleseed/renderer/kernel/lighting/backwardlightsampler.cpp
@@ -32,15 +32,9 @@
 
 // appleseed.renderer headers.
 #include "renderer/global/globallogger.h"
-#include "renderer/global/globaltypes.h"
-#include "renderer/kernel/intersection/intersector.h"
 #include "renderer/kernel/shading/shadingpoint.h"
-#include "renderer/modeling/edf/edf.h"
 #include "renderer/modeling/light/light.h"
-#include "renderer/modeling/object/iregion.h"
 #include "renderer/modeling/scene/scene.h"
-#include "renderer/modeling/shadergroup/shadergroup.h"
-#include "renderer/utility/triangle.h"
 
 // appleseed.foundation headers.
 #include "foundation/math/sampling/mappings.h"
@@ -62,7 +56,7 @@ namespace renderer
 BackwardLightSampler::BackwardLightSampler(
     const Scene&                        scene,
     const ParamArray&                   params)
-  : m_params(params)
+  : LightSamplerBase(params)
   , m_emitting_triangle_hash_table(m_triangle_key_hasher)
 {
     // Read which sampling algorithm should the sampler use.
@@ -218,293 +212,6 @@ Dictionary BackwardLightSampler::get_params_metadata()
     return metadata;
 }
 
-void BackwardLightSampler::collect_non_physical_lights(
-    const AssemblyInstanceContainer&    assembly_instances,
-    const TransformSequence&            parent_transform_seq)
-{
-    for (const_each<AssemblyInstanceContainer> i = assembly_instances; i; ++i)
-    {
-        // Retrieve the assembly instance.
-        const AssemblyInstance& assembly_instance = *i;
-
-        // Retrieve the assembly.
-        const Assembly& assembly = assembly_instance.get_assembly();
-
-        // Compute the cumulated transform sequence of this assembly instance.
-        TransformSequence cumulated_transform_seq =
-            assembly_instance.transform_sequence() * parent_transform_seq;
-        cumulated_transform_seq.prepare();
-
-        // Recurse into child assembly instances.
-        collect_non_physical_lights(
-            assembly.assembly_instances(),
-            cumulated_transform_seq);
-
-        // Collect lights from this assembly.
-        collect_non_physical_lights(
-            assembly,
-            cumulated_transform_seq);
-    }
-}
-
-void BackwardLightSampler::collect_non_physical_lights(
-    const Assembly&                     assembly,
-    const TransformSequence&            transform_sequence)
-{
-    for (const_each<LightContainer> i = assembly.lights(); i; ++i)
-    {
-        // Retrieve the light.
-        const Light& light = *i;
-
-        NonPhysicalLightInfo light_info;
-        light_info.m_transform_sequence = transform_sequence;
-        light_info.m_light = &light;
-
-        if (m_use_light_tree &&
-            ((light.get_flags() & Light::LightTreeCompatible) != 0))
-        {
-            // Copy the light into the light vector.
-            m_light_tree_lights.push_back(light_info);
-        }
-        else
-        {
-            // Copy the light into the light vector.
-            const size_t light_index = m_non_physical_lights.size();
-            m_non_physical_lights.push_back(light_info);
-
-            // Insert the light into the CDF.
-            // todo: compute importance.
-            float importance = 1.0f;
-            importance *= light.get_uncached_importance_multiplier();
-            m_non_physical_lights_cdf.insert(light_index, importance);
-        }
-    }
-}
-
-void BackwardLightSampler::collect_emitting_triangles(
-    const AssemblyInstanceContainer&    assembly_instances,
-    const TransformSequence&            parent_transform_seq)
-{
-    for (const_each<AssemblyInstanceContainer> i = assembly_instances; i; ++i)
-    {
-        // Retrieve the assembly instance.
-        const AssemblyInstance& assembly_instance = *i;
-
-        // Retrieve the assembly.
-        const Assembly& assembly = assembly_instance.get_assembly();
-
-        // Compute the cumulated transform sequence of this assembly instance.
-        TransformSequence cumulated_transform_seq =
-            assembly_instance.transform_sequence() * parent_transform_seq;
-        cumulated_transform_seq.prepare();
-
-        // Recurse into child assembly instances.
-        collect_emitting_triangles(
-            assembly.assembly_instances(),
-            cumulated_transform_seq);
-
-        // Collect emitting triangles from this assembly instance.
-        collect_emitting_triangles(
-            assembly,
-            assembly_instance,
-            cumulated_transform_seq);
-    }
-}
-
-void BackwardLightSampler::collect_emitting_triangles(
-    const Assembly&                     assembly,
-    const AssemblyInstance&             assembly_instance,
-    const TransformSequence&            transform_sequence)
-{
-    // Loop over the object instances of the assembly.
-    const size_t object_instance_count = assembly.object_instances().size();
-    for (size_t object_instance_index = 0; object_instance_index < object_instance_count; ++object_instance_index)
-    {
-        // Retrieve the object instance.
-        const ObjectInstance* object_instance = assembly.object_instances().get_by_index(object_instance_index);
-
-        // Retrieve the materials of the object instance.
-        const MaterialArray& front_materials = object_instance->get_front_materials();
-        const MaterialArray& back_materials = object_instance->get_back_materials();
-
-        // Skip object instances without light-emitting materials.
-        if (!has_emitting_materials(front_materials) && !has_emitting_materials(back_materials))
-            continue;
-
-        double object_area = 0.0;
-
-        // Compute the object space to world space transformation.
-        // todo: add support for moving light-emitters.
-        const Transformd& object_instance_transform = object_instance->get_transform();
-        const Transformd& assembly_instance_transform = transform_sequence.get_earliest_transform();
-        const Transformd global_transform = assembly_instance_transform * object_instance_transform;
-
-        // Retrieve the object.
-        Object& object = object_instance->get_object();
-
-        // Retrieve the region kit of the object.
-        Access<RegionKit> region_kit(&object.get_region_kit());
-
-        // Loop over the regions of the object.
-        const size_t region_count = region_kit->size();
-        for (size_t region_index = 0; region_index < region_count; ++region_index)
-        {
-            // Retrieve the region.
-            const IRegion* region = (*region_kit)[region_index];
-
-            // Retrieve the tessellation of the region.
-            Access<StaticTriangleTess> tess(&region->get_static_triangle_tess());
-
-            // Loop over the triangles of the region.
-            const size_t triangle_count = tess->m_primitives.size();
-            for (size_t triangle_index = 0; triangle_index < triangle_count; ++triangle_index)
-            {
-                // Fetch the triangle.
-                const Triangle& triangle = tess->m_primitives[triangle_index];
-
-                // Skip triangles without a material.
-                if (triangle.m_pa == Triangle::None)
-                    continue;
-
-                // Fetch the materials assigned to this triangle.
-                const size_t pa_index = static_cast<size_t>(triangle.m_pa);
-                const Material* front_material =
-                    pa_index < front_materials.size() ? front_materials[pa_index] : 0;
-                const Material* back_material =
-                    pa_index < back_materials.size() ? back_materials[pa_index] : 0;
-
-                // Skip triangles that don't emit light.
-                if ((front_material == 0 || front_material->has_emission() == false) &&
-                    (back_material == 0 || back_material->has_emission() == false))
-                    continue;
-
-                // Retrieve object instance space vertices of the triangle.
-                const GVector3& v0_os = tess->m_vertices[triangle.m_v0];
-                const GVector3& v1_os = tess->m_vertices[triangle.m_v1];
-                const GVector3& v2_os = tess->m_vertices[triangle.m_v2];
-
-                // Transform triangle vertices to assembly space.
-                const GVector3 v0_as = object_instance_transform.point_to_parent(v0_os);
-                const GVector3 v1_as = object_instance_transform.point_to_parent(v1_os);
-                const GVector3 v2_as = object_instance_transform.point_to_parent(v2_os);
-
-                // Compute the support plane of the hit triangle in assembly space.
-                const GTriangleType triangle_geometry(v0_as, v1_as, v2_as);
-                TriangleSupportPlaneType triangle_support_plane;
-                triangle_support_plane.initialize(TriangleType(triangle_geometry));
-
-                // Transform triangle vertices to world space.
-                const Vector3d v0(assembly_instance_transform.point_to_parent(v0_as));
-                const Vector3d v1(assembly_instance_transform.point_to_parent(v1_as));
-                const Vector3d v2(assembly_instance_transform.point_to_parent(v2_as));
-
-                // Compute the geometric normal to the triangle and the area of the triangle.
-                Vector3d geometric_normal = compute_triangle_normal(v0, v1, v2);
-                const double geometric_normal_norm = norm(geometric_normal);
-                if (geometric_normal_norm == 0.0)
-                    continue;
-                const double rcp_geometric_normal_norm = 1.0 / geometric_normal_norm;
-                const double rcp_area = 2.0 * rcp_geometric_normal_norm;
-                const double area = 0.5 * geometric_normal_norm;
-                geometric_normal *= rcp_geometric_normal_norm;
-                assert(is_normalized(geometric_normal));
-
-                if (object_instance->flip_normals())
-                    geometric_normal = -geometric_normal;
-
-                Vector3d n0, n1, n2;
-
-                if (triangle.m_n0 != Triangle::None &&
-                    triangle.m_n1 != Triangle::None &&
-                    triangle.m_n2 != Triangle::None)
-                {
-                    // Retrieve object instance space vertex normals.
-                    const Vector3d n0_os = Vector3d(tess->m_vertex_normals[triangle.m_n0]);
-                    const Vector3d n1_os = Vector3d(tess->m_vertex_normals[triangle.m_n1]);
-                    const Vector3d n2_os = Vector3d(tess->m_vertex_normals[triangle.m_n2]);
-
-                    // Transform vertex normals to world space.
-                    n0 = normalize(global_transform.normal_to_parent(n0_os));
-                    n1 = normalize(global_transform.normal_to_parent(n1_os));
-                    n2 = normalize(global_transform.normal_to_parent(n2_os));
-
-                    if (object_instance->flip_normals())
-                    {
-                        n0 = -n0;
-                        n1 = -n1;
-                        n2 = -n2;
-                    }
-                }
-                else
-                    n0 = n1 = n2 = geometric_normal;
-
-                for (size_t side = 0; side < 2; ++side)
-                {
-                    // Retrieve the material; skip sides without a material or without emission.
-                    const Material* material = side == 0 ? front_material : back_material;
-                    if (material == 0 || material->has_emission() == false)
-                        continue;
-
-                    // Accumulate the object area for OSL shaders.
-                    object_area += area;
-
-                    // Create a light-emitting triangle.
-                    EmittingTriangle emitting_triangle;
-                    emitting_triangle.m_assembly_instance = &assembly_instance;
-                    emitting_triangle.m_object_instance_index = object_instance_index;
-                    emitting_triangle.m_region_index = region_index;
-                    emitting_triangle.m_triangle_index = triangle_index;
-                    emitting_triangle.m_v0 = v0;
-                    emitting_triangle.m_v1 = v1;
-                    emitting_triangle.m_v2 = v2;
-                    emitting_triangle.m_n0 = side == 0 ? n0 : -n0;
-                    emitting_triangle.m_n1 = side == 0 ? n1 : -n1;
-                    emitting_triangle.m_n2 = side == 0 ? n2 : -n2;
-                    emitting_triangle.m_geometric_normal = side == 0 ? geometric_normal : -geometric_normal;
-                    emitting_triangle.m_triangle_support_plane = triangle_support_plane;
-                    emitting_triangle.m_area = static_cast<float>(area);
-                    emitting_triangle.m_rcp_area = static_cast<float>(rcp_area);
-                    emitting_triangle.m_triangle_prob = 0.0f;   // will be initialized once the emitting triangle CDF is built
-                    emitting_triangle.m_material = material;
-
-                    const size_t emitting_triangle_index = m_emitting_triangles.size();
-
-                    // Store the light-emitting triangle.
-                    m_emitting_triangles.push_back(emitting_triangle);
-
-                    // When using the CDF, insert the light-emitting triangle into the CDF.
-                    if (!m_use_light_tree)
-                    {
-                        // Retrieve the EDF and get the importance multiplier.
-                        float importance_multiplier = 1.0f;
-                        if (const EDF* edf = material->get_uncached_edf())
-                            importance_multiplier = edf->get_uncached_importance_multiplier();
-
-                        // Compute the probability density of this triangle.
-                        const float triangle_importance = m_params.m_importance_sampling ? static_cast<float>(area) : 1.0f;
-                        const float triangle_prob = triangle_importance * importance_multiplier;
-
-                        // Insert the light-emitting triangle into the CDF.
-                        m_emitting_triangles_cdf.insert(emitting_triangle_index, triangle_prob);
-                    }
-                }
-            }
-        }
-
-        store_object_area_in_shadergroups(
-            &assembly_instance,
-            object_instance,
-            static_cast<float>(object_area),
-            front_materials);
-
-        store_object_area_in_shadergroups(
-            &assembly_instance,
-            object_instance,
-            static_cast<float>(object_area),
-            back_materials);
-    }
-}
-
 void BackwardLightSampler::build_emitting_triangle_hash_table()
 {
     const size_t emitting_triangle_count = m_emitting_triangles.size();
@@ -634,35 +341,6 @@ void BackwardLightSampler::sample_emitting_triangle(
 
     // Compute the probability density of this sample.
     light_sample.m_probability = triangle_prob * emitting_triangle.m_rcp_area;
-}
-
-void BackwardLightSampler::store_object_area_in_shadergroups(
-    const AssemblyInstance*             assembly_instance,
-    const ObjectInstance*               object_instance,
-    const float                         object_area,
-    const MaterialArray&                materials)
-{
-    for (size_t i = 0, e = materials.size(); i < e; ++i)
-    {
-        if (const Material* m = materials[i])
-        {
-            if (const ShaderGroup* sg = m->get_uncached_osl_surface())
-            {
-                if (sg->has_emission())
-                    sg->set_surface_area(assembly_instance, object_instance, object_area);
-            }
-        }
-    }
-}
-
-
-//
-// BackwardLightSampler::Parameters class implementation.
-//
-
-BackwardLightSampler::Parameters::Parameters(const ParamArray& params)
-  : m_importance_sampling(params.get_optional<bool>("enable_importance_sampling", false))
-{
 }
 
 }   // namespace renderer
