+++ src/appleseed/renderer/kernel/lighting/forwardlightsampler.h
@@ -32,12 +32,13 @@
 
 // appleseed.renderer headers.
 #include "renderer/kernel/lighting/lightsample.h"
-#include "renderer/kernel/lighting/lighttree.h"
+#include "renderer/kernel/lighting/lightsamplerbase.h"
 #include "renderer/kernel/lighting/lighttypes.h"
 #include "renderer/kernel/shading/shadingray.h"
 
 // appleseed.foundation headers.
 #include "foundation/core/concepts/noncopyable.h"
+#include "foundation/math/cdf.h"
 #include "foundation/math/hash.h"
 #include "foundation/utility/containers/hashtable.h"
 
@@ -66,16 +67,13 @@ namespace renderer
 //
 
 class ForwardLightSampler
-  : public foundation::NonCopyable
+  : public LightSamplerBase
 {
   public:
     // Constructor.
     ForwardLightSampler(
         const Scene&                        scene,
         const ParamArray&                   params = ParamArray());
-
-    // Return the number of non-physical lights in the scene.
-    size_t get_non_physical_light_count() const;
 
     // Return the number of emitting triangles in the scene.
     size_t get_emitting_triangle_count() const;
@@ -96,50 +94,8 @@ class ForwardLightSampler
         LightSample&                        light_sample) const;
 
   private:
-    struct Parameters
-    {
-        const bool m_importance_sampling;
-
-        explicit Parameters(const ParamArray& params);
-    };
-
-    typedef std::vector<NonPhysicalLightInfo> NonPhysicalLightVector;
-    typedef std::vector<EmittingTriangle> EmittingTriangleVector;
-    typedef foundation::CDF<size_t, float> EmitterCDF;
-
-    const Parameters            m_params;
-
-    NonPhysicalLightVector      m_non_physical_lights;
-    size_t                      m_non_physical_light_count;
-
-    EmittingTriangleVector      m_emitting_triangles;
-
-    EmitterCDF                  m_non_physical_lights_cdf;
-    EmitterCDF                  m_emitting_triangles_cdf;
-
-    EmittingTriangleKeyHasher   m_triangle_key_hasher;
-    EmittingTriangleHashTable   m_emitting_triangle_hash_table;
-
-    // Recursively collect non-physical lights from a given set of assembly instances.
-    void collect_non_physical_lights(
-        const AssemblyInstanceContainer&    assembly_instances,
-        const TransformSequence&            parent_transform_seq);
-
-    // Collect non-physical lights from a given assembly.
-    void collect_non_physical_lights(
-        const Assembly&                     assembly,
-        const TransformSequence&            transform_sequence);
-
-    // Recursively collect emitting triangles from a given set of assembly instances.
-    void collect_emitting_triangles(
-        const AssemblyInstanceContainer&    assembly_instances,
-        const TransformSequence&            parent_transform_seq);
-
-    // Collect emitting triangles from a given assembly.
-    void collect_emitting_triangles(
-        const Assembly&                     assembly,
-        const AssemblyInstance&             assembly_instance,
-        const TransformSequence&            transform_sequence);
+    EmittingTriangleKeyHasher               m_triangle_key_hasher;
+    EmittingTriangleHashTable               m_emitting_triangle_hash_table;
 
     // Build a hash table that allows to find the emitting triangle at a given shading point.
     void build_emitting_triangle_hash_table();
@@ -170,23 +126,12 @@ class ForwardLightSampler
         const size_t                        triangle_index,
         const float                         triangle_prob,
         LightSample&                        sample) const;
-
-    void store_object_area_in_shadergroups(
-        const AssemblyInstance*             assembly_instance,
-        const ObjectInstance*               object_instance,
-        const float                         object_area,
-        const MaterialArray&                materials);
 };
 
 
 //
 // ForwardLightSampler class implementation.
 //
-
-inline size_t ForwardLightSampler::get_non_physical_light_count() const
-{
-    return m_non_physical_light_count;
-}
 
 inline size_t ForwardLightSampler::get_emitting_triangle_count() const
 {
